[Move] #11 Query 익스텐션을 Utility 디렉토리로 분리
